requirements.txt
streamlit
pyaudio
wave
google-cloud-speech